+++ essentials/src/components/CoreConcepts.jsx
@@ -1,0 +1,15 @@
+import CoreConcept from "./CoreConcept.jsx";
+import { CORE_CONCEPTS } from "../data.js";
+
+export default function CoreConcepts() {
+  return (
+    <section id="core-concept">
+      <h2>Core Concepts!</h2>
+      <ul>
+        {CORE_CONCEPTS.map((conceptItem) => (
+          <CoreConcept {...conceptItem} />
+        ))}
+      </ul>
+    </section>
+  );
+}
